show login error message & refactor file name
