+++ packages/ui/src/components/layout.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Content } from "rbx";
 
-import Navigation from "./Navigation";
+import Navigation from "./navigation";
 
-const Layout: React.FC<{ children: JSX.Element[] | JSX.Element | null }> = ({ children }) => (
+const Layout: React.FC<{ children: React.ReactNode }> = ({ children }) => (
     <Content>
         <Navigation />
         {children}
